requirements.txt
Django<=4.0.6
djangorestframework<=3.13.1
psycopg2-binary<=2.9.3
Pillow<=9.2.0
django-sendfile2<=0.6.1